// ang9-web/src/app/models/exam-detail.ts
export interface ExamDetail {
    Id:          string;
    ExamName:    string;
    Scheme:      string;
    Branch:      string;
    Semester:    string;
    SubjectCode: string;
    SubjectName: string;
    Date:        string;
    Time:        string;
    Status:      string;
    TotalMarks:  string;
    Attendance:  string;
    MarksStatus: string;
  }